V0.1.1

+++ src/main/java/com/kekecreations/configurable_falls/mixin/HayBlockMixin.java
@@ -16,6 +16,6 @@ public class HayBlockMixin {
 
     @Inject(method = "fallOn", at = @At("HEAD"), cancellable = true)
     public void fallOn(Level p_153362_, BlockState p_153363_, BlockPos p_153364_, Entity p_153365_, float p_153366_, CallbackInfo ci) {
-        p_153365_.causeFallDamage(p_153366_, ConfigurableFallsCommonConfigs.HAY_BALE_FALL_DAMAGE_PERCENTAGE.get(), p_153362_.damageSources().fall());
+        p_153365_.causeFallDamage(p_153366_, ConfigurableFallsCommonConfigs.HAY_BALE_FALL_DAMAGE_PERCENTAGE.get().floatValue(), p_153362_.damageSources().fall());
     }
 }
